Rename staff_handler package to staff.handler
